+++ react-todo-props/src/@context/components/TodoContainer.jsx
@@ -2,19 +2,17 @@ import {TodoInput} from "./TodoInput";
 import {TodoProvider} from "../hooks/useTodoContext";
 import {TodoList} from "./TodoList";
 import {useEffect, useState} from "react";
-import {LocalStorage} from "../../utils/LocalStorage";
+import {storage} from "../../domain/storage";
 
 export const TodoContainer = () => {
-    const storage = new LocalStorage();
+    const [todoList, setTodoList] = useState(storage.getItem('todoList', []));
 
-    const [todoList,setTodoList] = useState(storage.getItem('todoList',[]));
-
-    useEffect(()=>{
-        storage.setItem('todoList',todoList);
-    },[todoList]);
+    useEffect(() => {
+        storage.setItem('todoList', todoList);
+    }, [todoList]);
 
     return (
-        <TodoProvider value={{todoList,setTodoList}}>
+        <TodoProvider value={{todoList, setTodoList}}>
             <TodoInput/>
             <TodoList/>
         </TodoProvider>
