+++ react-todo-props/src/@context/components/TodoModify.jsx
@@ -1,6 +1,6 @@
 import {useState} from "react";
 import {useTodoContext} from "../hooks/useTodoContext";
-import {TodoService} from "../../utils/TodoService";
+import {TodoService} from "../../domain/TodoService";
 
 export const TodoModify = ({todo}) => {
     const {todoList, setTodoList} = useTodoContext();
